LNX-177: Add papaya & migrate page file ID from u64 to u32

+++ lnx-page-store/src/file/page_op_log/tests/reader.rs
@@ -9,7 +9,7 @@ use crate::file::{ctx, scheduler};
 use crate::layout::log;
 use crate::layout::log::{LogEntry, LogOp};
 
-const FILE_ID: u64 = 1;
+const FILE_ID: u32 = 1;
 
 #[rstest::rstest]
 #[trace]
